Step-4. CodeClimate remarks corrections. 1
