+++ src/main/java/ru/vk/itmo/tveritinalexandr/Loader.java
@@ -1,0 +1,48 @@
+package ru.vk.itmo.tveritinalexandr;
+
+import java.io.IOException;
+import java.lang.foreign.Arena;
+import java.lang.foreign.MemorySegment;
+import java.nio.channels.FileChannel;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.FileAlreadyExistsException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
+import java.nio.file.StandardOpenOption;
+import java.util.ArrayList;
+import java.util.List;
+
+public class Loader {
+    public static List<MemorySegment> loadOrRecover(Path storagePath, Arena arena) throws IOException {
+        Path indexTmp = storagePath.resolve("index.tmp");
+        Path indexFile = storagePath.resolve("index.idx");
+
+        if (Files.exists(indexTmp)) {
+            Files.move(indexTmp, indexFile, StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
+        } else {
+            try {
+                Files.createFile(indexFile);
+            } catch (FileAlreadyExistsException ignored) {
+                // it is ok, actually it is normal state
+            }
+        }
+
+        List<String> existedFiles = Files.readAllLines(indexFile, StandardCharsets.UTF_8);
+        List<MemorySegment> result = new ArrayList<>(existedFiles.size());
+        for (String fileName : existedFiles) {
+            Path file = storagePath.resolve(fileName);
+            try (FileChannel fileChannel = FileChannel.open(file, StandardOpenOption.READ, StandardOpenOption.WRITE)) {
+                MemorySegment fileSegment = fileChannel.map(
+                        FileChannel.MapMode.READ_WRITE,
+                        0,
+                        Files.size(file),
+                        arena
+                );
+                result.add(fileSegment);
+            }
+        }
+
+        return result;
+    }
+}
