+++ src/main/java/ru/vk/itmo/tveritinalexandr/MergeIterator.java
@@ -1,6 +1,10 @@
 package ru.vk.itmo.tveritinalexandr;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.Iterator;
+import java.util.NoSuchElementException;
+import java.util.PriorityQueue;
 
 public class MergeIterator<T> implements Iterator<T> {
