+++ src/main/java/ru/vk/itmo/tveritinalexandr/Utils.java
@@ -1,0 +1,29 @@
+package ru.vk.itmo.tveritinalexandr;
+
+import java.lang.foreign.MemorySegment;
+import java.lang.foreign.ValueLayout;
+
+public class Utils {
+
+    public static long recordsCount(MemorySegment segment) {
+        long indexSize = indexSize(segment);
+        return indexSize / Long.BYTES / 2;
+    }
+
+    public static long indexSize(MemorySegment segment) {
+        return segment.get(ValueLayout.JAVA_LONG_UNALIGNED, 0);
+    }
+
+    public static MemorySegment slice(MemorySegment page, long start, long end) {
+        return page.asSlice(start, end - start);
+    }
+
+    public static long tombstone(long offset) {
+        return 1L << 63 | offset;
+    }
+
+    public static long normalize(long value) {
+        return value & ~(1L << 63);
+    }
+
+}
